+++ dataobj/translator.cc
@@ -108,7 +108,7 @@ recode(const char *src,bool translate_from_utf,bool translate_to_utf)
 			else if(  translate_from_utf  ) {
 				// make latin from UTF8 (ignore overflows!)
 				int len=0;
-				*dst++ = c = utf8_to_utf16((const utf8*)src, &len);
+				*dst++ = c = (uint8)utf8_to_utf16((const utf8*)src, &len);
 				src += len;
 			}
 		}
@@ -224,9 +224,9 @@ DBG_DEBUG("translator::init_city_names()","file %p",file);
 DBG_MESSAGE("translator::init_city_names", "reading failed, creating random names.");
 		// Hajo: try to read list failed, create random names
 		for(int i=0; i<anz_t1; i++) {
-			const int l1 = strlen(translator::translate(name_t1[i]));
+			const long l1 = strlen(translator::translate(name_t1[i]));
 			for(int j=0; j<anz_t2; j++) {
-				const int l2 = strlen(translator::translate(name_t2[j]));
+				const long l2 = strlen(translator::translate(name_t2[j]));
 				char *c=(char *)guarded_malloc(l1+l2+1);
 				sprintf(c, "%s%s", translator::translate(name_t1[i]), translator::translate(name_t2[j]));
 				namen_liste.insert(c);
